JavadocTagInfo has been moved to javadoc package. It reduces number of cycle dependencies between packages. Part of issue #46
